chore: remove unused code and address pr comments about api vers
